New graphics but broken
New Features:
Added new graphics for heavy fire.

Current Bugs:
The graphics for fire only work once and don't loop.
The graphics activate on all heavy ships at the beginning of the game.
The direction of visible fire is not the same as the facing direction.
Ship models have disappeared.